Make Sprintf.format tokenize the format string at compile time
When the format string is a literal string, this approximately
doubles the speed. Additionally, it can now be used with a variable
number of arguments, or an array. (e.g. Sprintf.format("%s%s%s", "hi",
"hello", "howdy"))

Unfortunately, this implementation is slightly slower for non-
literal format strings, (this could be remedied by using original
code for non-literals) but the normal use case is an inline string.